+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v5.9-SNAPSHOT
 - Speedup in skipping the untouched Matchers.
 - Detection for Google Go, Google Docs, Google Docs Script
 - New class of Device and Agent: Voice
+- Added experimental filter for logstash
 
 v5.8
 ===